move 'deserialize' to child classes and eliminate cyclic dependencies
